Add rudimentary environment switching
This toggles between environments, similar to the GOV.UK toolkit:
https://github.com/alphagov/govuk-browser-extension

I plan on extending this to preserve the path in the URL when switching
environments when viewing a specific FutureLearn URL.

I had to update the manifest version to 3 in order for promises to work
with the tabs API on Chrome. This is not tested on Firefox yet.
See https://developer.chrome.com/docs/extensions/mv3/promises/